change data available test - left only verify and read file without rebuild check
clean before run
