+++ src/test/qa/data_availability_test.js
@@ -7,7 +7,6 @@ const P = require('../../util/promise');
 const promise_utils = require('../../util/promise_utils');
 const s3ops = require('../qa/s3ops');
 const af = require('../qa/functions/agent_functions');
-const api = require('../../api');
 
 require('../../util/dotenv').load();
 
@@ -139,57 +138,6 @@ function readFiles() {
         });
 }
 
-function getFileHealthStatus(key) {
-    let result = false;
-    let parts = [];
-    const rpc = api.new_rpc_default_only('wss://' + server_ip + ':8443');
-    const client = rpc.new_client({});
-    let auth_params = {
-        email: 'demo@noobaa.com',
-        password: 'DeMo1',
-        system: 'demo'
-    };
-    return client.create_auth_token(auth_params)
-        .then(() => P.resolve(client.object.read_object_mappings({
-            bucket,
-            key,
-            adminfo: true
-        })))
-        .then(res => {
-            parts = res.parts;
-            let chunkAvailable = parts.filter(chunk => chunk.chunk.adminfo.health === 'available').length;
-            let chunkNum = parts.length;
-            if (chunkAvailable === chunkNum) {
-                console.log('Available chunks number ' + chunkAvailable + ' all amount chunks ' + chunkNum);
-                result = true;
-            } else {
-                console.warn('Some chunk of file ' + key + ' has non available status');
-                result = false;
-            }
-        })
-        .catch(err => console.warn('Read chunk with error ' + err))
-        .then(() => result);
-}
-
-function waitForRebuildObjects(file) {
-    let retries = 0;
-    let rebuild = false;
-    console.log('Waiting for rebuild object ' + file);
-    return promise_utils.pwhile(
-        () => rebuild === false && retries !== 36,
-        () => P.resolve(getFileHealthStatus(file))
-            .then(res => {
-                if (res) {
-                    rebuild = res;
-                } else {
-                    retries += 1;
-                    console.log('Waiting for rebuild object' + file + ' - will wait for extra 5 seconds retries ' + retries);
-                }
-            })
-            .catch(e => console.warn('Waiting for rebuild file ' + file + 'with error ' + e))
-            .delay(5000));
-}
-
 function clean_up_dataset() {
     console.log('runing clean up files from bucket ' + bucket);
     return s3ops.get_list_files(server_ip, bucket, '')
@@ -204,21 +152,12 @@ function stopAgentsAndCheckFiles() {
         .then(res => {
             stopped_oses = res;
             //waiting for rebuild files by chunks and parts
-            return P.each(files, file => waitForRebuildObjects(file));
-        })
-        //Read and verify the read
-        .then(() => P.each(files, file => getFileHealthStatus(file)
-            .then(res => {
-                if (res === true) {
-                    console.log('File ' + file + ' rebuild successfully');
-                } else {
-                    saveErrorAndResume('File ' + file + ' didn\'t rebuild');
-                }
-            })))
-        .then(readFiles);
+            return readFiles();
+        });
 }
 
 return azf.authenticate()
+    .then(() => af.clean_agents(azf, osesSet, suffix))
     .then(() => af.createRandomAgents(azf, server_ip, storage, vnet, agents_number, suffix, osesSet))
     .then(res => {
         oses = res;
@@ -237,13 +176,16 @@ return azf.authenticate()
         failures_in_test = true;
         throw err;
     })
-    .finally(() => af.clean_agents(azf, oses, suffix)
-        .then(clean_up_dataset))
     .then(() => {
         if (failures_in_test) {
             console.error(':( :( Errors during data available test (replicas) ): ):' + errors);
             process.exit(1);
+        } else {
+            return af.clean_agents(azf, oses, suffix)
+                .then(clean_up_dataset)
+                .then(() => {
+                    console.log(':) :) :) data available test (replicas files) were successful! (: (: (:');
+                    process.exit(0);
+                });
         }
-        console.log(':) :) :) data available test (replicas files) were successful! (: (: (:');
-        process.exit(0);
     });
